Move current echo into only header, to reduce nesting
